Add (GPN:Beam) to the user-agent string of Python GCS SDKs. (#27211)
* Add (GPN:Beam) to the user-agent string of Python GCS SDKs.

For a valid user-agent string, the version number should be
separated with the product name by "/". In addition, the
requested string of "(GPN:Beam)" is considered a comment,
and could not be the start of the user-agent string.

* Add (GPN:Beam) to the user-agent string of Python GCS SDKs.

For a valid user-agent string, the version number should be
separated with the product name by "/". In addition, the
requested string of "(GPN:Beam)" is considered a comment,
and could not be the start of the user-agent string.